trying to clean up the stats mess
